search scryfall api on /api/search (#12)
* search scryfall api on search
When I search for a card, the server will hit the scryfall api search endpoint.
This is awesome, cause now we'll get all cards in their database.
Need to figure out a cache of sorts that persists between sessions since Scryfall has asked to do that. Not sure how to accomplish as of now.
I added a TODO to create a Search model that will help me compare results in our db to the results saved in the Search table. Not sure how that would work but it might help?
Essentially I want to check our cache (db) before I hit the Scryfall API. I'd save the different query parameters, total results, last time cached (which can be updated every 24 hours), and maybe the different unique ids returned (might be too granular).

Anyway. Progress! Love it

* fix create types